+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Checkbox/Checkbox/CheckboxPage.swift
@@ -47,7 +47,8 @@ private struct CheckboxDemo: View {
             Spacer()
             OUDSCheckbox(isOn: $configurationModel.indicatorState,
                          accessibilityLabel: "app_components_checkbox_hint_a11y".localized(), // No LocalizedStringKey inference in the component
-                         isError: configurationModel.isError)
+                         isError: configurationModel.isError,
+                         isReadOnly: configurationModel.isReadOnly)
                 .disabled(!configurationModel.enabled)
                 .accessibilityIdentifier(A11YIdentifiers.componentCheckbox)
             Spacer()